fix(build): move sanitizeReturnUrl out of +server.ts, fix stepper aria & csp byte cap
- sanitizeReturnUrl lived in +server.ts (SvelteKit rejects non-handler exports -> build failed); moved to $lib/server/auth/returnUrl.ts

- FormSteps: restore aria-label + aria-current on .step-button (E2E selectors + WAI stepper pattern) after the button refactor dropped them

- csp-report: measure real UTF-8 byte length (Buffer.byteLength) instead of UTF-16 string length
